Crud de Cliente y de Usuario

+++ src/main/java/ups/edu/ec/Sistema_Farmacia_S/Repositorio/UsuarioRepositorio.java
@@ -3,5 +3,6 @@ package ups.edu.ec.Sistema_Farmacia_S.Repositorio;
 import org.springframework.data.repository.CrudRepository;
 import ups.edu.ec.Sistema_Farmacia_S.Modelo.Usuario;
 
-public interface UsuarioRepositorio  extends CrudRepository<Usuario, Long> {
+public interface UsuarioRepositorio  extends CrudRepository<Usuario, Integer> {
+
 }
